Split package into subfolders
