+++ powerauth-fido2-demo/src/main/java/com/wultra/app/powerauth/fido2/demo/PowerauthFido2DemoApplication.java
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.wultra.app.powerauthfido2demo;
+package com.wultra.app.powerauth.fido2.demo;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
