+++ powerauth-fido2-demo/src/main/java/com/wultra/app/powerauth/fido2/demo/controller/HomeController.java
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.wultra.app.powerauthfido2demo.controller;
+package com.wultra.app.powerauth.fido2.demo.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
